refactor: monorepo path to service
