+++ .gitignore
@@ -24,9 +24,9 @@ go.work.sum
 # env file
 .env
 
-monorepo/*/tmp/
+services/*/tmp/
 /tmp
 /data
 
-monorepo/*/go.mod
-monorepo/*/go.sum
+services/*/go.mod
+services/*/go.sum
