--- monorepo/account/internal/core/services/factories.go
@@ -1,13 +1,0 @@
-package services
-
-import (
-	"concurrency-simulator/monorepo/account/utils"
-	"database/sql"
-)
-
-func NewAccountService(driver *sql.DB) *AccountService {
-	return &AccountService{
-		log:    utils.NewRequestLogger(),
-		driver: driver,
-	}
-}
