+++ services/account/controllers/factories.go
@@ -1,0 +1,19 @@
+package controllers
+
+import (
+	"concurrency-simulator/services/account/internal/core/services"
+	"concurrency-simulator/services/antifraud/utils"
+	"concurrency-simulator/services/shared"
+	"os"
+)
+
+func NewAccountController() *AccountController {
+	dbUrl := os.Getenv("DB_URL")
+
+	db := shared.NewPostgresSingleton(dbUrl)
+
+	return &AccountController{
+		logger:  utils.NewRequestLogger(),
+		service: services.NewAccountService(db),
+	}
+}
